requirements.txt
PyMuPDF
sentence-transformers
torch
llama-cpp-python
numpy
nltk
spacy
pandas
langdetect
deep-translator